chatbot/requirements.txt
flask==2.3.3
flask-cors==4.0.0
gunicorn==21.2.0
python-dotenv==1.0.0
pymongo==4.6.1
nltk==3.8.1
scikit-learn==1.3.2
numpy==1.26.2
pandas==2.1.3
python-dateutil==2.8.2
googletrans==3.1.0a0
arabic-reshaper==3.0.0
python-bidi==0.4.2
pytube==15.0.0 